Api Gateway fully working and refactored price variable - it should be a double

+++ src/main/java/com/siramiks/OrderService/entity/Order.java
@@ -43,7 +43,7 @@ public class Order {
   private LocalDateTime createdAt;
 
   @Column(name = "order_total_price")
-  private long orderPrice;
+  private double orderPrice;
 
   @Column(name = "order_total_quantity")
   private long orderQuantity;
